+++ examples/02_typescript/src/Item.tsx
@@ -1,14 +1,15 @@
 import React, { useState } from 'react';
 
-import { createAsync, useAsync } from 'react-hooks-fetch';
+import { prefetch, useFetch } from 'react-hooks-fetch';
 
 import DisplayData from './DisplayData';
 
-const initialResult = createAsync<{ data: { first_name: string } }>('https://reqres.in/api/users/1?delay=3');
+const initialId = '1';
+const initialResult = prefetch<{ data: { first_name: string } }>(`https://reqres.in/api/users/${initialId}?delay=3`);
 
 const Item: React.FC = () => {
-  const [id, setId] = useState('1');
-  const result = useAsync(initialResult);
+  const [id, setId] = useState(initialId);
+  const result = useFetch(initialResult);
   return (
     <div>
       User ID: <input value={id} onChange={e => setId(e.target.value)} />