+++ examples/03_noprefetch/src/Item.tsx
@@ -1,14 +1,14 @@
 import React, { useState } from 'react';
 
-import { createStatic, useAsync } from 'react-hooks-fetch';
+import { initialize, useFetch } from 'react-hooks-fetch';
 
 import DisplayData from './DisplayData';
 
-const initialResult = createStatic({ data: { first_name: '' } });
+const initialResult = initialize({ data: { first_name: '' } });
 
 const Item: React.FC = () => {
   const [id, setId] = useState('');
-  const result = useAsync(initialResult);
+  const result = useFetch(initialResult);
   return (
     <div>
       User ID: <input value={id} onChange={e => setId(e.target.value)} />