+++ src/index.d.ts
@@ -1,20 +1,18 @@
+import { Component, ReactNode } from 'react';
+
+type Fallback = ReactNode | ((error?: Error, retry?: () => void) => ReactNode);
+
+export class ErrorBoundary extends Component<{ fallback: Fallback }> {}
+
 type Input<Data> = string | Request | (() => Promise<Data>);
+type Result<Data> = { data: Data };
 
-export type CreateAsync = <Data>(input: Input<Data>) => {
-  data: Data;
-};
+export const prefetch: <Data>(input: Input<Data>) => Result<Data>;
 
-export type CreateStatic = <Data>(emptyData: Data) => {
-  data: Data;
-};
+export const initialize: <Data>(initialData: Data) => Result<Data>;
 
-export type UseAsync = <Data>(
-  initialResult: { data: Data }
-) => {
-  data: Data;
+export const useFetch: <Data>(
+  initialResult: Result<Data>
+) => Result<Data> & {
   refetch: (input: Input<Data>) => void;
 };
-
-export const createAsync: CreateAsync;
-export const createStatic: CreateStatic;
-export const useAsync: UseAsync;